Add titles and sequence numbers to chunks
Hoping that it will improve embeddings of the chunks.

Note that I did not change the maximum length of the chunk. The number
of tokens tends to be smaller than the number of symbols in the input so
perhaps it is not a problem.

On the other hand, I had to raise a TODO noticing that a block of common
text was labelled as CODE_BLOCK which raised a question of how should
chunk numbering work.